Update a database of tags

Examine the named database, and update it if required (see ‘Details’). Report an error if the file does not exist, or if the file does exist, but it is not an SQLite file containing tables named version, tagMapping and tags.

updateDatabase(dbname = getDatabaseName(), debug = 0)

Arguments

dbname

character value naming the database.

debug

integer value indicating a debugging level. Use 0 for silent operation, or any integer exceeding 1 to show some output during the action of this function.

Details

The following is a history of changes to the database schema.

2024-04-20: two changes were made to the database setup. First, the version element of the version table was changed from 1 to 2. Second, the level element of the tags table was renamed as index. Both changes are automatically handled by useDatabase(), which alters old files to the new format. All other functions of the package use the new format, so the system is backwards-compatible.

Author

Dan Kelley