Run an Shiny App for Tagging CTD Features

The tags are stored in a SQLite database, for ease of processing in R or other software systems. The analyst's name and the time of analysis is stored along with each time stamp, to facilitate combining the judgements made by multiple analysts. A SQLite database is used for this storage of tags.

ctdtag(
  path = ".",
  suffix = "cnv",
  dbprefix = "~/ctdtag",
  mapping = list(),
  plotHeight = 500,
  debug = 0
)

Arguments

path

character value naming the directory in which to search for CTD files.

suffix

character value indicating the file suffix that is taken to indicate CTD files. This is interpreted in a case-independent manner, so the default value of "cnv" would match both "station1.cnv" and "STATION2.CNV".

dbprefix

character value for the start of the name of the database file.

mapping

a list that connects numerical tag codes to character values indicating their meaning. For example, list("mixed-layer depth"=0) might be used in a study of mixed-layer depths. By default, mappings get set in a simple way, with numerical value 0 mapping to string "0", etc. The value of mapping is only used in creating a new database. After that, its value will be ignored if supplied, to avoid the confusion that would arise with multiple intended mappings. The mapping is stored in the database as a table named tagMapping, the contents of which are displayed in the table that appears below the plot.

plotHeight

numeric value for the height of the plot, in pixels.

debug

integer value indicating a debugging level. Use 0 for silent operation, or any integer exceeding 1 to show some output during the action of this function.

Details

Instructions are provided with the Help button (or by typing ?). Tagging information is stored in a sqlite3 database file, by default named ctdtag_USERNAME.db, where USERNAME is the login name of the analyst. If this file does not exist, it is created; otherwise, the existing tags (for the file undergoing analysis) are displayed on the plots, as a starting point.

Author

Dan Kelley